requirements.txt
rpi.gpio
rpimotorlib
